+++ src/kwset.h
@@ -23,7 +23,7 @@
 
 struct kwsmatch
 {
-  int index;			/* Index number of matching keyword. */
+  size_t index;			/* Index number of matching keyword. */
   size_t offset[1];		/* Offset of each submatch. */
   size_t size[1];		/* Length of each submatch. */
 };
